requirements.txt
cassandra-driver==3.25.0
certifi==2019.11.28
Click==7.0
colorhash==1.0.3
configparser==5.0.2
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
geomet==0.2.1.post1
gunicorn==20.1.0
imbalanced-learn==0.8.0
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.11.0
joblib==0.14.1
json5==0.9.6
jsonschema==3.2.0
kiwisolver==1.1.0
kneed==0.5.1
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
numpy==1.18.1
pandas==0.25.3
psutil==5.8.0
pycodestyle==2.8.0
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
regexp==0.1
scikit-learn==0.24.0
scipy==1.5.4
six==1.14.0
sklearn==0.0
SQLAlchemy==1.4.24
threadpoolctl==2.2.0
toml==0.10.2
tzlocal==2.1
Werkzeug==0.16.1
wincertstore==0.2
xgboost==0.90
zipp==2.0.1